+++ tests/test_printinfo.py
@@ -30,7 +30,7 @@ def test_versions(capsys):
     out2, _ = capsys.readouterr()
 
     # They have to be the same, except time (run at slightly different times)
-    assert out1b[75:] == out2[75:]
+    assert out1b[100:] == out2[100:]
 
     # Check the 'Pretty'/'plain'-version, providing a package as list
     out3 = versions('plain', add_pckg=[pip, ])
@@ -38,9 +38,9 @@ def test_versions(capsys):
     out3c = versions('Pretty', add_pckg=[pip, ])
 
     # They have to be the same, except time (run at slightly different times)
-    assert out3[75:] == out3b[75:]
+    assert out3[100:] == out3b[100:]
     if IPython:
-        assert out3[75:] == out3c.data[75:]
+        assert out3[100:] == out3c.data[100:]
     else:
         assert out3c is None